Add Better Handling of Unmapped Message Types
This can be a problem because there is a mismatch between the way the subscription filter is built and the way that message handler mappings are built.
Filters do not include the assembly name, but the handler mapping does.
This can present in the situation where a developer moves a type into a different assembly, whilst keeping its original namespace.