Update InvokeHandlerBehaviour to Unwrap and Re-Throw Exceptions from the Invoked Handler
InvokeHandlerBehaviour invokes HandleAsync method via reflection, which means that if the handler throws an exception, this will be wrapped in a TargetInvocationException.
This change allows behaviours within the processing pipeline to catch the original exception thrown from Handler whilst preserving the stack trace, and not have to be concerned with TargetInvocationException and therefore implementation details of how the handler was invoked.