Allow additive pipeline behaviours without recreating default pipeline
Instead of adding the "terminating" behaviours to the default pipeline, the terminating behaviours are added just-in-time as the PatLiteOptions are built. This allows consumers to request a default pipeline, then add custom behaviours they are interested in, without needing to recreate the default pipeline (and therefore understand the other behaviours and their order - including Pat's "internal" behaviours).